+++ src/types/tasks/Golem.cairo
@@ -7,7 +7,7 @@ pub impl Golem of TaskTrait {
     }
 
     #[inline]
-    fn description(count: u32) -> ByteArray {
+    fn description(count: u128) -> ByteArray {
         "Unlock Golem skin"
     }
 }
